Bug fixes for section issues.
